Move operator parser to op module
